requirements_undone.txt
h11==0.14.0
htmldate==1.4.3
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.13.3
idna==3.4
importlib-metadata==6.7.0
isodate==0.6.1
Jinja2==3.1.2
joblib==1.2.0
jusText==3.0.0
kor==0.6.1
langchain==0.0.209
langchainplus-sdk==0.0.17
langcodes==3.3.0
lit==16.0.6
llama-index==0.6.31
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
monotonic==1.6
mpmath==1.3.0
msal==1.22.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauth2client==4.1.3
olefile==0.46
openai==0.27.8
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
param==1.13.0
Pillow==9.5.0
pinecone-client==2.2.2
platformdirs==3.8.0
posthog==2.4.2
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.7
PyDrive==1.3.1
Pygments==2.14.0
pygtrie==2.5.0
PyJWT==2.7.0
PyMySQL==1.0.3
pypandoc==1.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
rdflib==6.3.2
regex==2023.3.23
requests==2.28.2
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.16
stability-sdk==0.8.1
starlette==0.26.1
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.3.3
tld==0.13
tokenizers==0.13.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
trafilatura==1.6.1
transformers==4.27.4
typing-inspect==0.8.0
typing_extensions==4.5.0
tzlocal==5.0.1
unstructured==0.5.9
uritemplate==4.1.1
urllib3==1.26.15
uvicorn==0.21.1
watchfiles==0.19.0
websockets==11.0
wrapt==1.14.1
XlsxWriter==3.0.9
yarl==1.8.2
zipp==3.15.0
zstandard==0.20.0